+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.4
+### Added
+- Comment service exposes an element query to query for comments
+- Element query supports the {% paginate %} tag
+
+### Fixed
+- Compatibility for Craft 3.0.20
+
 ## 1.0.3
 ### Added
 - Functionality to provide feedback to users after comment has been posted